feat: deduplicate log sanitization and expand sensitive keys (#248)
