+++ tsconfig.node.json
@@ -13,6 +13,7 @@
   "include": [
     "vite.config.ts",
     "electron",
-    "src/utils/translations.ts"
+    "src/utils/translations.ts",
+    "src/utils/logSanitizer.ts"
   ]
 }
